+++ pkg/plugin/plugin.go
@@ -2,50 +2,52 @@ package plugin
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
-	"github.com/grafana/grafana-plugin-sdk-go/backend/datasource"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/httpclient"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/instancemgmt"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/resource/httpadapter"
+	"github.com/grafana/sentry-datasource/pkg/errors"
+	"github.com/grafana/sentry-datasource/pkg/handlers"
+	"github.com/grafana/sentry-datasource/pkg/query"
 	"github.com/grafana/sentry-datasource/pkg/sentry"
+	"github.com/grafana/sentry-datasource/pkg/util"
 )
 
-type SentryPlugin struct {
-	sentryClient sentry.SentryClient
-}
+var (
+	_ backend.QueryDataHandler      = (*SentryDatasource)(nil)
+	_ backend.CheckHealthHandler    = (*SentryDatasource)(nil)
+	_ instancemgmt.InstanceDisposer = (*SentryDatasource)(nil)
+)
 
+const (
+	PluginID string = "grafana-sentry-datasource"
+)
+
+// SentryDatasource is a struct that represents the Sentry datasource.
 type SentryDatasource struct {
-	IM instancemgmt.InstanceManager
+	backend.CallResourceHandler
+	client sentry.SentryClient
 }
 
-func (ds *SentryDatasource) getDatasourceInstance(ctx context.Context, pluginCtx backend.PluginContext) (*SentryPlugin, error) {
-	s, err := ds.IM.Get(ctx, pluginCtx)
-	if err != nil {
-		return nil, err
-	}
-	return s.(*SentryPlugin), nil
-}
-
-func NewDatasource() datasource.ServeOpts {
-	im := datasource.NewInstanceManager(getInstance)
-	host := &SentryDatasource{
-		IM: im,
-	}
-	return datasource.ServeOpts{
-		CheckHealthHandler:  host,
-		QueryDataHandler:    host,
-		CallResourceHandler: httpadapter.New(host.getResourceRouter()),
+// NewDatasourceInstance creates an instance of the SentryDatasource. It is a helper
+// function that is mostly used for testing.
+func NewDatasourceInstance(sc *sentry.SentryClient) *SentryDatasource {
+	return &SentryDatasource{
+		client: *sc,
 	}
 }
 
-func getInstance(ctx context.Context, s backend.DataSourceInstanceSettings) (instancemgmt.Instance, error) {
+// NewDatasource creates an instance factory for the SentryDatasource. It is consumed by
+// the `datasource.Manage` function to create a new instance of the datasource.
+func NewDatasource(ctx context.Context, s backend.DataSourceInstanceSettings) (instancemgmt.Instance, error) {
 	settings, err := GetSettings(s)
 	if err != nil {
 		return nil, err
 	}
 
-	// we need this options to load the secure proxy configuration
+	// we need these options to load the secure proxy configuration
 	opt, err := s.HTTPClientOptions(ctx)
 	if err != nil {
 		return nil, err
@@ -60,7 +62,79 @@ func getInstance(ctx context.Context, s backend.DataSourceInstanceSettings) (ins
 	if err != nil {
 		return nil, err
 	}
-	return &SentryPlugin{
-		sentryClient: *sc,
+
+	ds := NewDatasourceInstance(sc)
+
+	// these are used to proxy requests to Sentry
+	ds.CallResourceHandler = httpadapter.New(ds.getResourceRouter())
+
+	return ds, nil
+}
+
+// Dispose is a callback that is called when the datasource is being disposed.
+func (ds *SentryDatasource) Dispose() {
+	// this is a no-op for now
+}
+
+// QueryData is the entrypoint for handling data queries from Grafana.
+func (ds *SentryDatasource) QueryData(ctx context.Context, req *backend.QueryDataRequest) (*backend.QueryDataResponse, error) {
+	// we don't pass a PluginContext while testing, so we need to check if it's nil
+	// before logging the datasource name to avoid a panic
+	if req.PluginContext.DataSourceInstanceSettings != nil {
+		backend.Logger.Debug("Query", "datasource", req.PluginContext.DataSourceInstanceSettings.Name)
+	}
+
+	response := backend.NewQueryDataResponse()
+
+	for _, q := range req.Queries {
+		res := createResponse(q, ds.client)
+		response.Responses[q.RefID] = res
+	}
+
+	return response, nil
+}
+
+// createResponse is a helper function that creates a response for a given query.
+func createResponse(backendQuery backend.DataQuery, client sentry.SentryClient) backend.DataResponse {
+	response := backend.DataResponse{}
+
+	query, err := query.GetQuery(backendQuery)
+	if err != nil {
+		return errors.GetErrorResponse(response, "", err)
+	}
+
+	switch query.QueryType {
+	case "issues":
+		return handlers.HandleIssues(client, query, backendQuery, response)
+	case "events":
+		return handlers.HandleEvents(client, query, backendQuery, response)
+	case "eventsStats":
+		return handlers.HandleEventsStats(client, query, backendQuery, response)
+	case "metrics":
+		return handlers.HandleMetrics(client, query, backendQuery, response)
+	case "statsV2":
+		return handlers.HandleStatsV2(client, query, backendQuery, response)
+	default:
+		response.Error = errors.ErrorUnknownQueryType
+		response.ErrorSource = backend.ErrorSourceDownstream
+	}
+
+	return response
+}
+
+// CheckHealth is a callback that is called when Grafana requests a health check for the datasource during setup.
+func (ds *SentryDatasource) CheckHealth(ctx context.Context, req *backend.CheckHealthRequest) (*backend.CheckHealthResult, error) {
+	projects, err := ds.client.GetProjects(ds.client.OrgSlug, false)
+	if err != nil {
+		errorMessage := err.Error()
+		return &backend.CheckHealthResult{
+			Status:  backend.HealthStatusError,
+			Message: errorMessage,
+		}, nil
+	}
+
+	return &backend.CheckHealthResult{
+		Status:  backend.HealthStatusOk,
+		Message: fmt.Sprintf("%s. %v projects found.", util.SuccessfulHealthCheckMessage, len(projects)),
 	}, nil
 }