+++ pkg/plugin/router.go
@@ -5,18 +5,24 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
-	"github.com/grafana/grafana-plugin-sdk-go/backend/resource/httpadapter"
 	"github.com/grafana/sentry-datasource/pkg/sentry"
 )
 
-func (host *SentryDatasource) getResourceRouter() *mux.Router {
+func (ds *SentryDatasource) getResourceRouter() *mux.Router {
 	router := mux.NewRouter()
-	router.HandleFunc("/api/0/organizations", host.withDatasourceHandler(GetOrganizationsHandler)).Methods("GET")
-	router.HandleFunc("/api/0/organizations/{organization_slug}/projects", host.withDatasourceHandler(GetProjectsHandler)).Methods("GET")
-	router.HandleFunc("/api/0/organizations/{organization_slug}/teams", host.withDatasourceHandler(GetOrganizationTeamsHandler)).Methods("GET")
-	router.HandleFunc("/api/0/teams/{organization_slug}/{team_slug}/projects", host.withDatasourceHandler(GetTeamsProjectsHandler)).Methods("GET")
-	router.NotFoundHandler = http.HandlerFunc(host.withDatasourceHandler(DefaultResourceHandler))
+	router.HandleFunc("/api/0/organizations", ds.withDatasourceHandler(GetOrganizationsHandler)).Methods("GET")
+	router.HandleFunc("/api/0/organizations/{organization_slug}/projects", ds.withDatasourceHandler(GetProjectsHandler)).Methods("GET")
+	router.HandleFunc("/api/0/organizations/{organization_slug}/teams", ds.withDatasourceHandler(GetOrganizationTeamsHandler)).Methods("GET")
+	router.HandleFunc("/api/0/teams/{organization_slug}/{team_slug}/projects", ds.withDatasourceHandler(GetTeamsProjectsHandler)).Methods("GET")
+	router.NotFoundHandler = http.HandlerFunc(ds.withDatasourceHandler(DefaultResourceHandler))
 	return router
+}
+
+func (ds *SentryDatasource) withDatasourceHandler(getHandler func(d *sentry.SentryClient) http.HandlerFunc) func(rw http.ResponseWriter, r *http.Request) {
+	return func(rw http.ResponseWriter, r *http.Request) {
+		h := getHandler(&ds.client)
+		h.ServeHTTP(rw, r)
+	}
 }
 
 func GetOrganizationsHandler(client *sentry.SentryClient) http.HandlerFunc {
@@ -62,7 +68,7 @@ func GetOrganizationTeamsHandler(client *sentry.SentryClient) http.HandlerFunc {
 			http.Error(rw, "invalid orgSlug", http.StatusBadRequest)
 			return
 		}
-		teams, err := client.ListOrganizationTeams(orgSlug)
+		teams, err := client.ListOrganizationTeams(orgSlug, true)
 		writeResponse(teams, err, rw)
 	}
 }
@@ -70,20 +76,6 @@ func GetOrganizationTeamsHandler(client *sentry.SentryClient) http.HandlerFunc {
 func DefaultResourceHandler(client *sentry.SentryClient) http.HandlerFunc {
 	return func(rw http.ResponseWriter, r *http.Request) {
 		http.Error(rw, "not a valid resource call", http.StatusNotImplemented)
-	}
-}
-
-func (host *SentryDatasource) withDatasourceHandler(getHandler func(d *sentry.SentryClient) http.HandlerFunc) func(rw http.ResponseWriter, r *http.Request) {
-	return func(rw http.ResponseWriter, r *http.Request) {
-		ctx := r.Context()
-		pluginContext := httpadapter.PluginConfigFromContext(ctx)
-		datasource, err := host.getDatasourceInstance(ctx, pluginContext)
-		if err != nil {
-			http.Error(rw, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		h := getHandler(&datasource.sentryClient)
-		h.ServeHTTP(rw, r)
 	}
 }